Introduce ability to preserve newline state
Today if an error causes a `Print` struct to be dropped, there's no way to know later if the next line needs to be empty or not. This PR introduces a global writer that tracks the number of newlines previously written to the output. Calling `Print::global` tells the writer to use that prior information when resuming.