Introduce global print functions (#21)
* Refactor write styling logic out of `Print` struct

This moves the bulk of write formatting out of the Print struct state machine. This opens up the ability to re-use it with a global writer and retain consistent formatting.

* Introduce ability to preserve newline state

Today if an error causes a `Print` struct to be dropped, there's no way to know later if the next line needs to be empty or not. This PR introduces a global writer that tracks the number of newlines previously written to the output. Calling `Print::global` tells the writer to use that prior information when resuming.

* Expose global write functions

Details in the module docs and style guide

* PR URL

* Clippy

* Flush after write

* Consistent newlines in style guide

* Prefix function api to make sub-bullet functions clearer

* Emit logging format in  docs from functions

* Update readme block languages