Refactor shared "announce" code (#27)
* .gitignore

* Refactor shared "announce" code

The prior implementation relied on an internal trait to give identical implementations of the "announcement" states i.e. error/warning/important to Bullet and SubBullet. This was always a little difficult to work with. Since the refactor of this logic into the shared `write.rs` module, we can instead duplicate the code and rely on the internal interface to share instead. This is more ergonomic.

This commit is a pure refactor, the end user should see no change.