Fix version specifier (#30)
I mistakenly thought 0.4 behaved similar to ruby's `~> 0.4` which would allow 0.4 >= and < 1.0. However that's not the case. It WOULD be the case if we we're above 1.0


From https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html:

```
1.2.3  :=  >=1.2.3, <2.0.0
1.2    :=  >=1.2.0, <2.0.0
1      :=  >=1.0.0, <2.0.0
```

However, because fun_run is pre-1.0 the rules are surprisingly different:


```
0.2.3  :=  >=0.2.3, <0.3.0
0.2    :=  >=0.2.0, <0.3.0
0.0.3  :=  >=0.0.3, <0.0.4
0.0    :=  >=0.0.0, <0.1.0
0      :=  >=0.0.0, <1.0.0
```

This version specifier is what I intended when I wrote `0.4` thinking it would be like the `1.2` example above.