Introduce `global::with_locked_writer`
New function `global::with_locked_writer` is introduced to allow consistently capturing write output. This function is designed for use in testing output or in other non-reentrant capture cases. This blocks all threads using this function but one from executing so that a deterministic and consistent output is captured. Previously tests could be written with a thread_local writer, however there's a subtle race condition in that approach if the output relies on "paragraph" style text.

The cause of that race condition is because `global::set_writer` not only mutates the writer, it also resets the trailing newlines count, which affects the output. This replacement for that behavior should be safer.