Ensure `set_writer` is not called inside of `with_locked_writer`
Doing this with different types will already cause a panic at the point of the box downcast. Adding an explicit panic gives us two things: It forces the backtrace to appear at the point where the problem is introduced. It also guarantees we don't silently not record something as is the case if the `set_writer` is accidentally set to the same type.